# tests/snapshots/openapi.json
{
  "components": {
    "schemas": {
...
      "ApprovalRequest": {
        "properties": {
          "approval_comment": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Approval Comment"
          }
        },
        "title": "ApprovalRequest",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "QualityAnalyticsResponse": {
        "properties": {
          "entity_type_coverage": {
            "additionalProperties": {
              "type": "integer"
            },
            "title": "Entity Type Coverage",
            "type": "object"
          },
          "grade": {
            "title": "Grade",
            "type": "string"
          },
          "metrics": {
            "$ref": "#/components/schemas/QualityMetrics"
          },
          "overall_score": {
            "title": "Overall Score",
            "type": "number"
          },
          "property_fill_rates": {
            "additionalProperties": {
              "type": "number"
            },
            "title": "Property Fill Rates",
            "type": "object"
          },
          "quality_gate_status": {
            "title": "Quality Gate Status",
            "type": "string"
          },
          "requires_review": {
            "title": "Requires Review",
            "type": "boolean"
          },
          "transform_id": {
            "title": "Transform Id",
            "type": "string"
          },
          "violations": {
            "$ref": "#/components/schemas/QualityViolationsAnalytics"
          }
        },
        "required": [
          "transform_id",
          "overall_score",
          "grade",
          "quality_gate_status",
          "requires_review",
          "metrics",
          "violations",
          "property_fill_rates",
          "entity_type_coverage"
        ],
        "title": "QualityAnalyticsResponse",
        "type": "object"
      },
...
      "QualityRuleType": {
        "description": "Types of quality rules that can be applied.",
        "enum": [
          "format",
          "business",
          "cross_entity",
          "distribution",
          "consistency"
        ],
        "title": "QualityRuleType",
        "type": "string"
      },
      "QualitySeverity": {
        "description": "Severity levels for quality violations.",
        "enum": [
          "error",
          "warning",
          "info"
        ],
        "title": "QualitySeverity",
        "type": "string"
      },
...
      "RejectQualityRequest": {
        "properties": {
          "rejection_reason": {
            "title": "Rejection Reason",
            "type": "string"
          }
        },
        "required": [
          "rejection_reason"
        ],
        "title": "RejectQualityRequest",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "Graphora API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/quality/analytics/{transform_id}": {
      "get": {
        "description": "Return aggregated analytics for a transform.",
        "operationId": "get_quality_analytics_api_v1_quality_analytics__transform_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "transform_id",
            "required": true,
            "schema": {
              "title": "Transform Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QualityAnalyticsResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Quality Analytics",
        "tags": [
          "quality"
        ]
      }
    },
    "/api/v1/quality/approve/{transform_id}": {
      "post": {
        "description": "User approves quality results and proceeds to merge.",
        "operationId": "approve_quality_results_api_v1_quality_approve__transform_id__post",
        "parameters": [
          {
            "in": "path",
            "name": "transform_id",
            "required": true,
            "schema": {
              "title": "Transform Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApprovalRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Approve Quality Results",
        "tags": [
          "quality"
        ]
      }
    },
    "/api/v1/quality/export/{transform_id}": {
      "get": {
        "description": "Export violations as CSV for a given transform.",
        "operationId": "export_quality_violations_api_v1_quality_export__transform_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "transform_id",
            "required": true,
            "schema": {
              "title": "Transform Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Export Quality Violations",
        "tags": [
          "quality"
        ]
      }
    },
...
    "/api/v1/quality/reject/{transform_id}": {
      "post": {
        "description": "User rejects quality results - stops the process.",
        "operationId": "reject_quality_results_api_v1_quality_reject__transform_id__post",
        "parameters": [
          {
            "in": "path",
            "name": "transform_id",
            "required": true,
            "schema": {
              "title": "Transform Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RejectQualityRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Reject Quality Results",
        "tags": [
          "quality"
        ]
      }
    },
    "/api/v1/quality/results/{transform_id}": {
      "delete": {
        "description": "Delete quality results for a transform.",
        "operationId": "delete_quality_results_api_v1_quality_results__transform_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "transform_id",
            "required": true,
            "schema": {
              "title": "Transform Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Delete Quality Results",
        "tags": [
          "quality"
        ]
      },
      "get": {
        "description": "Get quality validation results for a transform.",
        "operationId": "get_quality_results_api_v1_quality_results__transform_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "transform_id",
            "required": true,
            "schema": {
              "title": "Transform Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QualityResults"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Quality Results",
        "tags": [
          "quality"
        ]
      }
    },
    "/api/v1/quality/summary": {
      "get": {
        "description": "Return recent quality summaries for the authenticated user.",
        "operationId": "get_quality_summary_endpoint_api_v1_quality_summary_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 10,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/QualitySummaryItem"
                  },
                  "title": "Response Get Quality Summary Endpoint Api V1 Quality Summary Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Quality Summary Endpoint",
        "tags": [
          "quality"
        ]
      }
    },
    "/api/v1/quality/violations/{transform_id}": {
      "get": {
        "description": "List persisted quality violations for a transform with optional filtering.",
        "operationId": "list_quality_violations_api_v1_quality_violations__transform_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "transform_id",
            "required": true,
            "schema": {
              "title": "Transform Id",
              "type": "string"
            }
          },
          {
            "description": "Filter by quality rule type",
            "in": "query",
            "name": "violation_type",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/QualityRuleType"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by quality rule type",
              "title": "Violation Type"
            }
          },
          {
            "description": "Filter by violation severity",
            "in": "query",
            "name": "severity",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/QualitySeverity"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by violation severity",
              "title": "Severity"
            }
          },
          {
            "description": "Filter by entity type",
            "in": "query",
            "name": "entity_type",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by entity type",
              "title": "Entity Type"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 100,
              "maximum": 500,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/QualityViolation"
                  },
                  "title": "Response List Quality Violations Api V1 Quality Violations  Transform Id  Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Quality Violations",
        "tags": [
          "quality"
        ]
      }
    },
...
  }
}